requirements.txt
mlflow
boto3
scikit-learn
pandas
dvc
dvc-s3
pillow
numpy